refactor: rename req4att to credential
