+++ packages/core/src/__integrationtests__/Attestation.spec.ts
@@ -23,7 +23,7 @@ import { FullDidDetails } from '@kiltprotocol/did'
 import { Crypto } from '@kiltprotocol/utils'
 import * as Attestation from '../attestation'
 import { getRemoveTx, getRevokeTx } from '../attestation/Attestation.chain'
-import * as Credential from '../requestforattestation'
+import * as Credential from '../credential'
 import { disconnect } from '../kilt'
 import * as Claim from '../claim'
 import * as CType from '../ctype'
