+++ packages/core/src/__integrationtests__/Deposit.spec.ts
@@ -38,7 +38,7 @@ import {
 import { Balance } from '../balance'
 import * as Attestation from '../attestation'
 import * as Claim from '../claim'
-import * as Credential from '../requestforattestation'
+import * as Credential from '../credential'
 import { disconnect } from '../kilt'
 import { queryRaw } from '../attestation/Attestation.chain'
 import * as CType from '../ctype'
