+++ packages/core/src/credential/index.ts
@@ -5,4 +5,4 @@
  * found in the LICENSE file in the root directory of this source tree.
  */
 
-export * from './RequestForAttestation.js'
+export * from './Credential.js'
