+++ packages/core/src/index.ts
@@ -12,7 +12,7 @@
 import '@polkadot/api-augment'
 
 export * as Attestation from './attestation/index.js'
-export * as Credential from './requestforattestation/index.js'
+export * as Credential from './credential/index.js'
 export { Balance, BalanceUtils } from './balance/index.js'
 export * as Claim from './claim/index.js'
 export * as CType from './ctype/index.js'
